+++ CHANGELOG.md
@@ -5,6 +5,8 @@
     take effect until the "Publish" button is clicked
   * Splash image settings are now editable even when the MiniReel is
     published
+  * The last splash theme/ratio the user selected is now stored in
+    LocalStorage/the user DB
   * Generated Wordpress Shortcode has been updated to the latest format
     based on feedback from USAToday